separare files for constants and reducers
