+++ src/actionCreators.js
@@ -1,5 +1,5 @@
 import db, { firebase, provider } from './firebase';
-import { POSTDARE, ACCEPTDARE, DECLINEDARE } from './index';
+import { POSTDARE, ACCEPTDARE, DECLINEDARE } from './constants';
 
 export function signIn(user) {
   return {
