+++ src/constants.js
@@ -1,0 +1,3 @@
+export const POSTDARE = 'POSTDARE';
+export const ACCEPTDARE = 'ACCEPTDARE';
+export const DECLINEDARE = 'DECLINEDARE';
